+++ src/quill-cursors/cursor.ts
@@ -34,6 +34,8 @@ export default class Cursor {
   public build(options: IQuillCursorsOptions): HTMLElement {
     const element = document.createElement(Cursor.CONTAINER_ELEMENT_TAG);
     element.classList.add(Cursor.CURSOR_CLASS);
+    options.cursorCssClass && element.classList.add(options.cursorCssClass);
+    element.id = `ql-cursor-${ this.id }`;
     element.innerHTML = options.template;
     const selectionElement = element.getElementsByClassName(Cursor.SELECTION_CLASS)[0] as HTMLElement;
     const caretContainerElement = element.getElementsByClassName(Cursor.CARET_CONTAINER_CLASS)[0] as HTMLElement;