+++ src/quill-cursors/i-quill-cursors-options.ts
@@ -1,5 +1,6 @@
 export default interface IQuillCursorsOptions {
   template?: string;
+  cursorCssClass?: string;
   selectionChangeSource?: string;
   hideDelayMs?: number;
   hideSpeedMs?: number;